requirements.txt
google-generativeai
pandasql 
streamlit
python-dotenv
glob2 